Added new MicroVE instructions. Bug fixes.

+++ src/MicroVASMDotNET/Compilers/Specifications/ISpecTypes.cs
@@ -2,7 +2,7 @@
 {
     public interface ISpecTypes : ICompilerSpec
     {
-        bool GetTypeSize(string name, out byte size);
+        bool GetTypeSize(string name, out int size);
         bool GetTypeValue(string name, out ValueType type);
     }
 }
